mcp_server/requirements.txt
fastmcp
httpx
pydantic
python-dotenv
google-api-python-client
google-auth
